Todoリストの過去のTodo(完了済み)の非表示と並び替えを行い、カスタムフックにまとめた
