+++ client/src/components/organisms/study/TodoList.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable no-undef */
 import { FC, memo } from "react";
 import {
   Flex,
@@ -17,10 +16,10 @@ import {
 } from "@chakra-ui/react";
 import { AddIcon } from "@chakra-ui/icons";
 import type { ApolloError } from "@apollo/client";
-import { isWithinInterval, isToday, isBefore, addDays } from "date-fns";
 
 import { TodoWithCheck } from "../../molucules/todos/TodoWithCheck";
 import type { TodoData } from "../../../types/types";
+import { useTodoList } from "../../../hooks/study/useTodoList";
 
 type Props = {
   todos: Array<TodoData | null>;
@@ -34,83 +33,12 @@ type Props = {
 const tabs = ["全て", "今日", "期限切れ"] as const; //as const をつけてreadonlyにする
 
 /**
- * Todoの配列の中身がnullでないことをチェックする.
- *
- * @remarks trueなら、引数の配列の中身はTodoData型になる
- * @param todos Todoの配列
- * @returns Todoの配列の中身がnullでないか
- */
-export const isNonNullTodoData = (
-  todoDataList: Array<TodoData | null>,
-): todoDataList is Array<TodoData> => {
-  return todoDataList.length > 0;
-};
-
-/**
  * Todoリストを表示するコンポーネント.
  */
 export const TodoList: FC<Props> = memo((props) => {
   const { todos, loading, error, openReadModal, openAddModal } = props;
 
-  /**
-   * Todoをタブのタイプに応じてフィルタリングする.
-   *
-   * @returns フィルタリングしたtodoの配列
-   */
-  const getFilteredTodos = (tabType: "全て" | "今日" | "期限切れ") => {
-    const today = new Date();
-
-    // todosの中身がnullかどうかで型ガード
-    if (!isNonNullTodoData(todos)) {
-      return [];
-    }
-
-    // switch文で場合分け
-    switch (tabType) {
-      // 全ての場合
-      case "全て":
-        return todos;
-
-      case "今日":
-        // 今日の場合
-        return todos.filter((todo) => {
-          // 期間に今日が含まれているものを返す
-          const startDate = new Date(todo.startedAt);
-          if (todo?.finishedAt) {
-            // 複数日間のタスク
-            const endDate = new Date(todo?.finishedAt);
-
-            // date-fnsのisWithinIntervalメソッドで、範囲内に入っているかどうかを判定
-            return (
-              isWithinInterval(today, {
-                start: startDate,
-                end: endDate,
-              }) ||
-              isToday(startDate) ||
-              isToday(endDate)
-            );
-          } else {
-            // 一日のタスク
-            // date-fnsのisTodayメソッドで、今日かどうかを判定
-            return isToday(startDate);
-          }
-        });
-      case "期限切れ":
-        // 期限切れの場合
-        return todos.filter((todo) => {
-          // 期限切れのもの
-          const startDate = new Date(todo?.startedAt);
-          if (todo?.finishedAt) {
-            // 複数日間のタスク
-            const endDate = new Date(todo?.finishedAt);
-            return isBefore(addDays(endDate, 1), today);
-          }
-          return isBefore(addDays(startDate, 1), today);
-        });
-      default:
-        return [];
-    }
-  };
+  const { getFilteredTodos } = useTodoList(todos);
 
   return (
     <>
